+++ src/main/scala/org/ergoplatform/settings/ChainSettings.scala
@@ -50,20 +50,4 @@ case class ChainSettings(protocolVersion: Byte,
 }
 
 
-case class ReemissionSettings(checkReemissionRules: Boolean,
-                              emissionNftId: ModifierId,
-                              reemissionTokenId: ModifierId,
-                              reemissionNftId: ModifierId,
-                              activationHeight: Int,
-                              reemissionStartHeight: Int,
-                              injectionBoxBytesEncoded: ModifierId) {
 
-  val emissionNftIdBytes: Array[Byte] = Algos.decode(emissionNftId).get
-  val reemissionNftIdBytes: Array[Byte] = Algos.decode(reemissionNftId).get
-  val reemissionTokenIdBytes: Array[Byte] = Algos.decode(reemissionTokenId).get
-
-  lazy val InjectionBoxBytes: Array[Byte] = Base16.decode(injectionBoxBytesEncoded).get
-  lazy val injectionBox: ErgoBox = ErgoBoxSerializer.parseBytes(InjectionBoxBytes)
-
-  val reemissionRules = new ReemissionRules(reemissionSettings = this)
-}
